+++ src/test/lit-element_styling_test.ts
@@ -541,6 +541,155 @@ suite('Static get styles', () => {
     assert.equal(getComputedStyleValue(level3!, 'border-top-width').trim(), '3px');
     assert.equal(getComputedStyleValue(level4!, 'border-top-width').trim(), '4px');
   });
+
+  test('`styles` can be a static field', async () => {
+    const name = generateElementName();
+    customElements.define(name, class extends LitElement {
+      static styles = [
+        css`div {
+          border: 2px solid blue;
+        }`,
+        css`span {
+          display: block;
+          border: 3px solid blue;
+        }`
+      ];
+
+      render() {
+        return htmlWithStyles`
+        <div>Testing1</div>
+        <span>Testing2</span>`;
+      }
+    });
+    const el = document.createElement(name);
+    container.appendChild(el);
+    await (el as LitElement).updateComplete;
+    const div = el.shadowRoot!.querySelector('div');
+    assert.equal(getComputedStyleValue(div!, 'border-top-width').trim(), '2px');
+    const span = el.shadowRoot!.querySelector('span');
+    assert.equal(getComputedStyleValue(span!, 'border-top-width').trim(),
+                 '3px');
+  });
+
+  test('can extend and augment `styles`', async () => {
+    const base = generateElementName();
+    customElements.define(base, class extends LitElement {
+      static get styles() { return css`div {
+          border: 2px solid blue;
+        }`;
+      }
+
+      render() {
+        return htmlWithStyles`
+        <div>Testing1</div>`;
+      }
+    });
+
+    const sub = generateElementName();
+    customElements.define(sub, class extends customElements.get(base) {
+      static get styles() {
+        return [
+          super.styles,
+          css`span {
+            display: block;
+            border: 3px solid blue;
+          }`
+        ];
+      }
+
+      render() {
+        return htmlWithStyles`
+        ${super.render()}
+        <span>Testing2</span>`;
+      }
+    });
+
+    const subsub = generateElementName();
+    customElements.define(subsub, class extends customElements.get(sub) {
+      static get styles() {
+        return [
+          super.styles,
+          css`p {
+            display: block;
+            border: 4px solid blue;
+          }`
+        ];
+      }
+
+      render() {
+        return htmlWithStyles`
+        ${super.render()}
+        <p>Testing3</p>`;
+      }
+    });
+    let el = document.createElement(base);
+    container.appendChild(el);
+    await (el as LitElement).updateComplete;
+    const div = el.shadowRoot!.querySelector('div');
+    assert.equal(getComputedStyleValue(div!, 'border-top-width').trim(), '2px');
+    el = document.createElement(sub);
+    container.appendChild(el);
+    await (el as LitElement).updateComplete;
+    const span = el.shadowRoot!.querySelector('span');
+    assert.equal(getComputedStyleValue(span!, 'border-top-width').trim(),
+                 '3px');
+    el = document.createElement(subsub);
+    container.appendChild(el);
+    await (el as LitElement).updateComplete;
+    const p = el.shadowRoot!.querySelector('p');
+    assert.equal(getComputedStyleValue(p!, 'border-top-width').trim(),
+                 '4px');
+  });
+
+  test('can extend and override `styles`', async () => {
+    const base = generateElementName();
+    customElements.define(base, class extends LitElement {
+      static get styles() { return css`div {
+          border: 2px solid blue;
+        }`;
+      }
+
+      render() {
+        return htmlWithStyles`
+        <div>Testing1</div>`;
+      }
+    });
+
+    const sub = generateElementName();
+    customElements.define(sub, class extends customElements.get(base) {
+      static get styles() { return css`div {
+          border: 3px solid blue;
+        }`;
+      }
+
+    });
+
+    const subsub = generateElementName();
+    customElements.define(subsub, class extends customElements.get(sub) {
+      static get styles() { return css`div {
+          border: 4px solid blue;
+        }`;
+      }
+
+    });
+    let el = document.createElement(base);
+    container.appendChild(el);
+    await (el as LitElement).updateComplete;
+    let div = el.shadowRoot!.querySelector('div');
+    assert.equal(getComputedStyleValue(div!, 'border-top-width').trim(), '2px');
+    el = document.createElement(sub);
+    container.appendChild(el);
+    await (el as LitElement).updateComplete;
+    div = el.shadowRoot!.querySelector('div');
+    assert.equal(getComputedStyleValue(div!, 'border-top-width').trim(),
+                 '3px');
+    el = document.createElement(subsub);
+    container.appendChild(el);
+    await (el as LitElement).updateComplete;
+    div = el.shadowRoot!.querySelector('div');
+    assert.equal(getComputedStyleValue(div!, 'border-top-width').trim(),
+                 '4px');
+  });
 });
 
 suite('ShadyDOM', () => {